+++ mplgrid/__init__.py
@@ -1,3 +1,3 @@
 """ This module imports the mplgrid functions"""
-from .mplgrid import grid_dimensions, grid
-from ._version import __version__
+from .mplgrid import grid, grid_dimensions
+from .__about__ import __version__