+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+:rocket: Version 0.2.1
+----------------------
+
+Fixes
+- Fix docstrings for aspect (height / width)
+
+:rocket: Version 0.2
+--------------------
+
+Changed
+- removed the ability to create endnote and title axes.
+- grid now returns axes and figure instead of a dictonary.
+- grid uses matplotlib plt.subplots instead of manually adding the axes with fig.add_axes()
+- aspect is flipped to match matplotlib (height / width) instead of the previous (width / height)
+- arguments are now renamed: grid_height to height, grid_width to width, and max_grid to max_side.
+
 :rocket: Version 0.1.0
 ----------------------